+++ templates/types/streaming/fastapi/app/api/routers/models.py
